Move organization related components under People module
