+++ module/People/config/module.config.php
@@ -61,21 +61,21 @@ return array(
 			'People\Assertion\MemberOfOrganizationAssertion' => 'People\Assertion\MemberOfOrganizationAssertion',
 		),
 	),
-	// 	'doctrine' => array(
-// 		'driver' => array(
-// 			 __NAMESPACE__ . '_driver' => array(
-// 				'class' => 'Doctrine\ORM\Mapping\Driver\AnnotationDriver',
-// 				'cache' => 'array',
-// 				'paths' => array(__DIR__ . '/../src/'. __NAMESPACE__ . '/Entity')
-// 			),
-// 			'orm_default' => array(
-// 				'drivers' => array(
-// 					__NAMESPACE__ . '\Entity' =>  __NAMESPACE__ . '_driver'
-// 				)
-// 			)
-// 		)
-// 	),
-// 	'listeners' => array(
-// 		'Application\OrganizationCommandsListener'
-// 	),	
+	'doctrine' => array(
+		'driver' => array(
+			 __NAMESPACE__ . '_driver' => array(
+				'class' => 'Doctrine\ORM\Mapping\Driver\AnnotationDriver',
+				'cache' => 'array',
+				'paths' => array(__DIR__ . '/../src/'. __NAMESPACE__ . '/Entity')
+			),
+			'orm_default' => array(
+				'drivers' => array(
+					__NAMESPACE__ . '\Entity' =>  __NAMESPACE__ . '_driver'
+				)
+			)
+		)
+	),
+	'listeners' => array(
+		'People\OrganizationCommandsListener'
+	),	
 );
