+++ module/TaskManagement/test/unit/TaskManagement/Controller/StreamsControllerTest.php
@@ -4,8 +4,8 @@ namespace TaskManagement\Controller;
 use ZFX\Test\Controller\ControllerTest;
 use Rhumsaa\Uuid\Uuid;
 use Application\Entity\User;
-use Application\Organization;
-use Application\Service\OrganizationService;
+use People\Organization;
+use People\Service\OrganizationService;
 use TaskManagement\Service\StreamService;
 use TaskManagement\Task;
 use TaskManagement\Stream;
