+++ tests/sql/init.sql
@@ -43,10 +43,10 @@ INSERT INTO users (id, status, createdAt, mostRecentEditAt, firstname, lastname,
 
 # organization 00000000-0000-0000-1000-000000000000
 INSERT INTO event_stream (eventId, version, eventName, payload, occurredOn, aggregate_type, aggregate_id) VALUES
-('907dd600-1e37-4e35-8045-7abfc5f60895',1,'Application\\OrganizationCreated','a:2:{s:2:"by";s:36:"60000000-0000-0000-0000-000000000000";s:12:"aggregate_id";s:36:"00000000-0000-0000-1000-000000000000";}', '2015-03-06T19:42:58.000000+0100','Application\\Organization','00000000-0000-0000-1000-000000000000'),
-('dbc7ecb0-63b5-40c9-8ae4-09c06449a4ce',2,'Application\\OrganizationUpdated','a:3:{s:4:"name";s:11:"O.R.A. Team";s:2:"by";s:36:"60000000-0000-0000-0000-000000000000";s:12:"aggregate_id";s:36:"00000000-0000-0000-1000-000000000000";}','2015-03-06T19:42:58.000000+0100','Application\\Organization','00000000-0000-0000-1000-000000000000'),
-('74548e90-569c-4e1b-958f-5b644243210c',3,'Application\\OrganizationMemberAdded','a:4:{s:6:"userId";s:36:"60000000-0000-0000-0000-000000000000";s:4:"role";s:5:"admin";s:2:"by";s:36:"60000000-0000-0000-0000-000000000000";s:12:"aggregate_id";s:36:"00000000-0000-0000-1000-000000000000";}','2015-03-11T13:35:32.000000+0100','Application\\Organization','00000000-0000-0000-1000-000000000000'),
-('75548e90-569c-4e1b-958f-5b644243210c',4,'Application\\OrganizationAccountChanged','a:3:{s:9:"accountId";s:36:"dcde992b-5aa9-4447-98ae-c8115906dcb7";s:2:"by";s:36:"60000000-0000-0000-0000-000000000000";s:12:"aggregate_id";s:36:"00000000-0000-0000-1000-000000000000";}','2015-03-11T13:35:32.000000+0100','Application\\Organization','00000000-0000-0000-1000-000000000000');
+('907dd600-1e37-4e35-8045-7abfc5f60895',1,'People\\OrganizationCreated','a:2:{s:2:"by";s:36:"60000000-0000-0000-0000-000000000000";s:12:"aggregate_id";s:36:"00000000-0000-0000-1000-000000000000";}', '2015-03-06T19:42:58.000000+0100','People\\Organization','00000000-0000-0000-1000-000000000000'),
+('dbc7ecb0-63b5-40c9-8ae4-09c06449a4ce',2,'People\\OrganizationUpdated','a:3:{s:4:"name";s:11:"O.R.A. Team";s:2:"by";s:36:"60000000-0000-0000-0000-000000000000";s:12:"aggregate_id";s:36:"00000000-0000-0000-1000-000000000000";}','2015-03-06T19:42:58.000000+0100','People\\Organization','00000000-0000-0000-1000-000000000000'),
+('74548e90-569c-4e1b-958f-5b644243210c',3,'People\\OrganizationMemberAdded','a:4:{s:6:"userId";s:36:"60000000-0000-0000-0000-000000000000";s:4:"role";s:5:"admin";s:2:"by";s:36:"60000000-0000-0000-0000-000000000000";s:12:"aggregate_id";s:36:"00000000-0000-0000-1000-000000000000";}','2015-03-11T13:35:32.000000+0100','People\\Organization','00000000-0000-0000-1000-000000000000'),
+('75548e90-569c-4e1b-958f-5b644243210c',4,'People\\OrganizationAccountChanged','a:3:{s:9:"accountId";s:36:"dcde992b-5aa9-4447-98ae-c8115906dcb7";s:2:"by";s:36:"60000000-0000-0000-0000-000000000000";s:12:"aggregate_id";s:36:"00000000-0000-0000-1000-000000000000";}','2015-03-11T13:35:32.000000+0100','People\\Organization','00000000-0000-0000-1000-000000000000');
 INSERT INTO `organizations` (id, name, createdAt, mostRecentEditAt, createdBy_id, mostRecentEditBy_id) VALUES
 ('00000000-0000-0000-1000-000000000000', 'O.R.A. Team','2014-11-06 13:11:05','2014-11-06 13:11:05', '60000000-0000-0000-0000-000000000000', '60000000-0000-0000-0000-000000000000');
 INSERT INTO organization_members(member_id, organization_id, role, createdAt, mostRecentEditAt, createdBy_id, mostRecentEditBy_id) VALUES
